lots of stuff, including code for indexeddb

+++ app/scripts/services/authService.js
@@ -36,20 +36,6 @@
                     localStorage.removeItem(localFiles.previousLogin);
                 }
 
-                if (true) { //offline -> localStorage/localDb-based login //!window.navigator.onLine
-
-                    //todo: check credentials against local saved ones, which are based on the last succesful logins.
-                    var sessionData = {
-                        loggedOnline: false,
-                        userName: loginData.login
-                    };
-
-                    sessionService.setSession(sessionData);
-                    deferred.resolve(sessionData);
-
-                    return deferred.promise;
-                }
-                else {
                     var data = 'grant_type=password&username=' + loginData.login + '&password=' + loginData.password;
 
                     $http.post(config.authTokenRoot + '/token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } })
@@ -63,16 +49,29 @@
                             };
 
                             sessionService.setSession(sessionData);
-
                             deferred.resolve(response);
 
                         }).error(function (err) {
-                            sessionService.destroy();
-                            deferred.reject(err);
+
+                            if (false) { //offline -> localStorage/localDb-based login //!window.navigator.onLine
+                                //todo: check credentials against local saved ones, which are based on the last succesful logins.
+                                var sessionData = {
+                                    loggedOnline: false,
+                                    userName: loginData.login
+                                };
+
+                                sessionService.setSession(sessionData);
+                                deferred.resolve(sessionData);
+
+                                return deferred.promise;
+                            }
+                            else {
+                                sessionService.destroy();
+                                deferred.reject(err);
+                            }
                         });
 
                     return deferred.promise;
-                }
             }
         }]);
 })();
